src/plugins/analysis/architecture_detection/requirements.txt
pyelftools
pyyaml
